refactor(sql): move to pg package
