+++ pkg/sql/pg/config.go
@@ -1,4 +1,4 @@
-package sql
+package pg
 
 import (
 	"github.com/kelseyhightower/envconfig"
@@ -14,6 +14,6 @@ func NewConfig() (*Config, error) {
 
 // Config for SQL.
 type Config struct {
-	AppName       string `envconfig:"APP_NAME" required:"true"`
-	PostgreSQLURL string `envconfig:"POSTGRESQL_URL" required:"true"`
+	AppName string `envconfig:"APP_NAME" required:"true"`
+	URL     string `envconfig:"POSTGRESQL_URL" required:"true"`
 }
